pyproject.toml
[project]
name = "takehome"
version = "0.1.0"
description = "User API with referral system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
]
